Add PerList.js component
